pyproject.toml
[tool.poetry]
name = "flask-api"
version = "0.1.0"
description = ""
authors = ["jozoshu"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2 = "^2.9.1"
pytz = "^2021.3"
bcrypt = "^3.2.0"
PyJWT = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
